requirements.txt
langgraph>=0.0.40
openai>=1.0.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
pydantic>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0